+++ database/migrations/2026_05_17_102920_create_history_produks.php
@@ -11,16 +11,18 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('history_produks', function (Blueprint $table) {
-            $table->id('id_history');
-            $table->unsignedBigInteger('id_produk');
-            $table->decimal('harga_reguler_lama', 10, 2)->nullable();
-            $table->decimal('harga_reguler_baru', 10, 2);
-            $table->decimal('harga_member_lama', 10, 2)->nullable();
-            $table->decimal('harga_member_baru', 10, 2);
-            $table->foreignId('id_admin_cabang')->constrained('admin_cabangs', 'id_admin_cabang')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('history_produks')) {
+            Schema::create('history_produks', function (Blueprint $table) {
+                $table->id('id_history');
+                $table->unsignedBigInteger('id_produk');
+                $table->decimal('harga_reguler_lama', 10, 2)->nullable();
+                $table->decimal('harga_reguler_baru', 10, 2);
+                $table->decimal('harga_member_lama', 10, 2)->nullable();
+                $table->decimal('harga_member_baru', 10, 2);
+                $table->foreignId('id_admin_cabang')->constrained('admin_cabangs', 'id_admin_cabang')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -28,6 +30,6 @@ return new class extends Migration
      */
     public function down(): void
     {
-        //
+        Schema::dropIfExists('history_produks');
     }
 };
